chore(ui): Extract signup steps into components
